Tests for GuestService and PollService

+++ src/main/java/no/hvl/dat250/h2020/group5/service/PollService.java
@@ -1,14 +1,13 @@
 package no.hvl.dat250.h2020.group5.service;
 
-import no.hvl.dat250.h2020.group5.dao.PollRepository;
-import no.hvl.dat250.h2020.group5.dao.UserRepository;
-import no.hvl.dat250.h2020.group5.dao.VoteRepository;
 import no.hvl.dat250.h2020.group5.entities.Poll;
+import no.hvl.dat250.h2020.group5.entities.User;
 import no.hvl.dat250.h2020.group5.entities.Vote;
 import no.hvl.dat250.h2020.group5.enums.AnswerType;
 import no.hvl.dat250.h2020.group5.enums.PollVisibilityType;
-import no.hvl.dat250.h2020.group5.entities.User;
-
+import no.hvl.dat250.h2020.group5.repositories.PollRepository;
+import no.hvl.dat250.h2020.group5.repositories.UserRepository;
+import no.hvl.dat250.h2020.group5.repositories.VoteRepository;
 import no.hvl.dat250.h2020.group5.responses.VotesResponse;
 import org.springframework.stereotype.Service;
 
@@ -46,14 +45,16 @@ public class PollService {
         }
     }
 
-    public boolean deletePoll(Long pollId) {
+    public boolean deletePoll(Long pollId, Long userId) {
         Optional<Poll> poll = pollRepository.findById(pollId);
         if (poll.isEmpty()) {
             return false;
-        } else {
+        }
+        if (isOwnerOrAdmin(poll.get(), userId)) {
             pollRepository.delete(poll.get());
             return true;
         }
+        return false;
     }
 
     public List<Poll> getAllPublicPolls() {
@@ -66,7 +67,7 @@ public class PollService {
 
     public List<Poll> getUserPolls(Long userId) {
         Optional<User> user = userRepository.findById(userId);
-        return user.map(value -> pollRepository.findAllByPollOwner(value)).orElse(null);
+        return user.map(pollRepository::findAllByPollOwner).orElse(null);
     }
 
     public Poll getPoll(Long pollId) {
@@ -86,7 +87,7 @@ public class PollService {
         return true;
     }
 
-    public boolean getPollStatus(Long pollId) {
+    public boolean isActivated(Long pollId) {
         Optional<Poll> poll = pollRepository.findById(pollId);
         if (poll.isEmpty()) {
             return false;
@@ -95,7 +96,7 @@ public class PollService {
         Instant startTime = poll.get().getStartTime().toInstant();
         Instant startTimePlusDuration = startTime.plusSeconds(poll.get().getPollDuration());
 
-        return !Instant.now().isAfter(startTimePlusDuration);
+        return Instant.now().isBefore(startTimePlusDuration);
     }
 
     public VotesResponse getNumberOfVotes(Long pollId) {
@@ -119,5 +120,15 @@ public class PollService {
         votesResponse.setNo(no);
         votesResponse.setYes(yes);
         return votesResponse;
+    }
+
+    private boolean isOwnerOrAdmin(Poll poll, Long userId) {
+        Optional<User> maybeUser = userRepository.findById(userId);
+        return maybeUser
+                .filter(
+                        user ->
+                                user.getId().equals(poll.getPollOwner().getId())
+                                        || user.getIsAdmin())
+                .isPresent();
     }
 }
